+++ web-app/resources/js/Pages/Public/Comment/Create.vue
@@ -1,31 +1,23 @@
 <template>
-  <Head :title="$t('comments.showTitle')"/>
+  <Head :title="$t('public.comments.createTitle')"/>
   <v-container fluid>
     <flash-messages/>
     <v-form @submit.prevent="submit">
       <v-row>
         <v-col cols="12" md="4">
-          <div :class="['text-h5', 'pa-2']">{{ $t('comments.showTitle') }}</div>
+          <div :class="['text-h5', 'pa-2']">{{ $t('public.comments.createTitle') }}</div>
         </v-col>
       </v-row>
       <v-row>
         <v-col cols="12" md="4">
-          <v-select
-            v-model="form.status_id"
-            :label="$t('users.fields.role')"
-            :items="statuses"
-            :item-title="item => $t('commentStatus.'+item.name)"
-            :rules="rules.required"
-            item-value="id"
-          ></v-select>
           <v-text-field
             v-model="form.name"
-            :label="$t('comments.fields.name')"
+            :label="$t('public.comments.fields.name')"
             :rules="rules.required"
           ></v-text-field>
           <v-textarea
             v-model="form.remarks"
-            :label="$t('comments.fields.remarks')"
+            :label="$t('public.comments.fields.remark')"
           ></v-textarea>
         </v-col>
       </v-row>
@@ -39,12 +31,12 @@
             elevation="0"
           > {{ $t('form.saveBtn') }}
           </v-btn>
-          <Link :href="$route('vegetation.show', {vegetation: vegetation.uuid})">
+          <Link :href="$route('public.vegetation.show', {vegetation: vegetation.uuid})">
             <v-btn
               prepend-icon="mdi-keyboard-return"
               size="large"
               class="ml-5"
-              :href="$route('vegetation.show', {vegetation: vegetation.uuid})"
+              :href="$route('public.vegetation.show', {vegetation: vegetation.uuid})"
               elevation="0"
             > {{ $t('form.backBtn') }}
             </v-btn>
@@ -62,11 +54,14 @@ import {required, email} from '@vee-validate/rules';
 import {useI18n} from "vue-i18n";
 import FlashMessages from "../../../Shared/FlashMessages.vue";
 
-const props = defineProps(['comment', 'vegetation', 'statuses']);
+const props = defineProps(['vegetation']);
 
 const {t} = useI18n({});
 
-const form = useForm(props.comment);
+const form = useForm({
+  name: null,
+  remarks: null
+});
 
 const rules = {
   required: [
@@ -78,8 +73,7 @@ async function submit(event) {
   const results = await event;
 
   if (results.valid) {
-    form.put(route('comments.update', {
-      comment: props.comment.uuid,
+    form.post(route('public.vegetation.comment.store', {
       vegetation: props.vegetation.uuid
     }));
   }