+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2025-04-24
+
+### Added
+
+- public profile
+- volunteer role
+
+## [0.6.0] - 2025-04-22
+
+### Added
+
+- public comments/mutations
+
 ## [0.5.1] - 2025-04-22
 
 ### Changed